refactored models into separe .py files
