+++ backend/models/sample.py
@@ -1,41 +1,28 @@
 from pydantic import BaseModel
 from typing import Union, Literal
 
+from .graph import SimpleGraphElement
+
+"""
+Here we define the models for the samples
+and the data associated with them.
+"""
 
 class Pangolin(BaseModel):
+    #Pangolin classification of a sample
     conflict: Union[int, str]
     type: str
     pangolearn_version: str
 
 class Variant(BaseModel):
+    # Variant detection of a sample
     dp: int
     aa: str
     id: str
     alt_freq: float
 
-class DashBoardGraphElement(BaseModel):
-    date: str
-    pango_count: int
-    pangolin: str
-
-class SimpleGraphElement(BaseModel):
-    date: str
-    count: int
-
-
-
-class GeneralStats(BaseModel):
-    passed_qc_samples: int
-    unique_pangos: int
-    unique_mutations: int
-
-class DashboardGraph(BaseModel):
-    dashboard_data: list[DashBoardGraphElement]
-    general_stats: GeneralStats
-    selection_criterions: list[str]
-
-
 class Sample(BaseModel):
+    # SarsCov2 sample
     _id: dict
     variants : Union[list[Variant],None] = None
     qc: dict
@@ -53,30 +40,8 @@ class Sample(BaseModel):
     sex: Union[str, None]
     Ct: Union[str, None]
     similar_samples: Union[list, None]
-    
 
 class GroupedSamples(BaseModel):
     samples: list[Sample]
     graph: list[SimpleGraphElement]
-class Basic_User(BaseModel):
-    username: str 
-    email: str
-    fullname: str
-    disabled: bool = False
-    scope: Literal['user','admin']
 
-class User(Basic_User):
-    password: str
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    username: Union[str, None] = None
-
-
-class UserInDB(User):
-    hashed_password: str
